Fix compiler warnings and Ultralight x64 crash
- Fix C4267/C4244: cast strlen/sizeof/ptrdiff_t to unsigned int in String.inl, StringReader, StringWriter, TypeDefinitionMacros, TypeVariable.inl
- Fix C26439: add noexcept to Json::Value swap specialization
- Fix C26495: default-initialize mData, mPayload in DynamicArrayC, GraphNode, GraphEdge
- Fix C6262: heap-allocate MainProcessingUnit in Main.cpp (was ~60KB on stack)
- Fix C4996: suppress GetVersionEx deprecation warning in CallStack.cpp
- Fix D9035: remove MinimalRebuild (/Gm) from 8 project files
- Suppress LNK4006 duplicate symbol noise from Ultralight SDK libs (/IGNORE:4006)
- Deploy Ultralight resources/ folder in DiaUIUltralight PreBuildEvent for x64

+++ Dia/DiaCore/Strings/String.inl
@@ -19,7 +19,7 @@ namespace Dia
 				DIA_ASSERT(str.length() <= Capacity(), "String being read in will not fit in size of string being allocated");
 
 				char* pointerToDst = reinterpret_cast<char*>(instance.Pointee());
-				Dia::Core::StringCopy(pointerToDst, &str[0], str.length());
+				Dia::Core::StringCopy(pointerToDst, &str[0], static_cast<unsigned int>(str.length()));
 			}
 
 			//------------------------------------------------------------------------------------
@@ -228,7 +228,7 @@ namespace Dia
 			template <unsigned int size> inline
 			String<size>& String<size>::Append( const char* s )
 			{
-				return Append(s, strlen(s));
+				return Append(s, static_cast<unsigned int>(strlen(s)));
 			}
 			
 			//-----------------------------------------------------------------------------
@@ -576,7 +576,7 @@ namespace Dia
 			template <unsigned int size> inline
 			unsigned int String<size>::Length() const
 			{
-				return strlen(&mData[0]);
+				return static_cast<unsigned int>(strlen(&mData[0]));
 			}
 
 			//-----------------------------------------------------------------------------
